refactor(packages): moved levels + level views into com.example.demo.level
