+++ src/javadocs/com.example.demo/com/example/demo/BossLevelView.html
@@ -85,8 +85,8 @@ loadScripts(document, 'script');</script>
 <h1 title="Class BossLevelView" class="title">Class BossLevelView</h1>
 </div>
 <div class="inheritance" title="Inheritance Tree"><a href="https://docs.oracle.com/en/java/javase/22/docs/api/java.base/java/lang/Object.html" title="class or interface in java.lang" class="external-link">java.lang.Object</a>
-<div class="inheritance"><a href="LevelView.html" title="class in com.example.demo">com.example.demo.LevelView</a>
-<div class="inheritance">com.example.demo.BossLevelView</div>
+<div class="inheritance"><a href="LevelView.html" title="class in com.example.demo">com.example.demo.level.LevelView</a>
+<div class="inheritance">com.example.demo.level.BossLevelView</div>
 </div>
 </div>
 <section class="class-description" id="class-description">
