+++ src/javadocs/com.example.demo/com/example/demo/LevelParent.html
@@ -86,7 +86,7 @@ loadScripts(document, 'script');</script>
 </div>
 <div class="inheritance" title="Inheritance Tree"><a href="https://docs.oracle.com/en/java/javase/22/docs/api/java.base/java/lang/Object.html" title="class or interface in java.lang" class="external-link">java.lang.Object</a>
 <div class="inheritance"><a href="https://docs.oracle.com/en/java/javase/22/docs/api/java.base/java/util/Observable.html" title="class or interface in java.util" class="external-link">java.util.Observable</a>
-<div class="inheritance">com.example.demo.LevelParent</div>
+<div class="inheritance">com.example.demo.level.LevelParent</div>
 </div>
 </div>
 <section class="class-description" id="class-description">
