+++ src/javadocs/com.example.demo/com/example/demo/LevelView.html
@@ -85,7 +85,7 @@ loadScripts(document, 'script');</script>
 <h1 title="Class LevelView" class="title">Class LevelView</h1>
 </div>
 <div class="inheritance" title="Inheritance Tree"><a href="https://docs.oracle.com/en/java/javase/22/docs/api/java.base/java/lang/Object.html" title="class or interface in java.lang" class="external-link">java.lang.Object</a>
-<div class="inheritance">com.example.demo.LevelView</div>
+<div class="inheritance">com.example.demo.level.LevelView</div>
 </div>
 <section class="class-description" id="class-description">
 <div class="horizontal-scroll">
