+++ src/main/java/com/example/demo/controller/Controller.java
@@ -9,11 +9,11 @@ import javafx.scene.Scene;
 import javafx.scene.control.Alert;
 import javafx.scene.control.Alert.AlertType;
 import javafx.stage.Stage;
-import com.example.demo.LevelParent;
+import com.example.demo.level.LevelParent;
 
 public class Controller implements Observer {
 
-	private static final String LEVEL_ONE_CLASS_NAME = "com.example.demo.LevelOne";
+	private static final String LEVEL_ONE_CLASS_NAME = "com.example.demo.level.LevelOne";
 	private final Stage stage;
 
 	public Controller(Stage stage) {
@@ -29,7 +29,7 @@ public class Controller implements Observer {
 
 	public void startLevelOne() {
 		try {
-			goToLevel("com.example.demo.LevelOne");
+			goToLevel("com.example.demo.level.LevelOne");
 		} catch (ClassNotFoundException | NoSuchMethodException | SecurityException | InstantiationException
 				 | IllegalAccessException | IllegalArgumentException | InvocationTargetException e) {
 			e.printStackTrace();
