+++ src/main/java/com/example/demo/level/LevelOne.java
@@ -1,9 +1,12 @@
-package com.example.demo;
+package com.example.demo.level;
+
+import com.example.demo.ActiveActorDestructible;
+import com.example.demo.EnemyPlane;
 
 public class LevelOne extends LevelParent {
 	
 	private static final String BACKGROUND_IMAGE_NAME = "/com/example/demo/images/background1.jpg";
-	private static final String NEXT_LEVEL = "com.example.demo.LevelTwo";
+	private static final String NEXT_LEVEL = "com.example.demo.level.LevelTwo";
 	private static final int TOTAL_ENEMIES = 5;
 	private static final int KILLS_TO_ADVANCE = 10;
 	private static final double ENEMY_SPAWN_PROBABILITY = .20;
