+++ src/main/java/com/example/demo/level/LevelView.java
@@ -1,5 +1,8 @@
-package com.example.demo;
+package com.example.demo.level;
 
+import com.example.demo.GameOverImage;
+import com.example.demo.HeartDisplay;
+import com.example.demo.WinImage;
 import javafx.scene.Group;
 
 public class LevelView {
