+++ src/main/java/module-info.java
@@ -8,4 +8,6 @@ module com.example.demo {
 
     opens com.example.demo to javafx.fxml;
     exports com.example.demo.controller;
+    opens level to javafx.fxml;
+    opens com.example.demo.level to javafx.fxml;
 }
